remove unit test flake - maps aren't ordered
